updating files and sh script

+++ features/files.sh
@@ -1,6 +1,9 @@
 #! /bin/sh
 
-rm automation.log
+if [ -f "automation.log" ]; then
+    rm automation.log
+fi
+
 #Check if exists path and remove all files generated for allure report
 if [ -d "features/reports/temp" ]
 then
